ml/cold_start/requirements.txt
# Cold Start Collaborative Filtering — Python Dependencies
# --------------------------------------------------------

# Core ML
scikit-learn>=1.4.0
numpy>=1.26.0
joblib>=1.3.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0

# Database
supabase>=2.3.0
postgrest>=0.16.0

# HTTP client (for DKT service + supabase SDK)
httpx>=0.27.0

# Logging
loguru>=0.7.0

# Config
python-dotenv>=1.0.0

# Scheduling (for retrain cron)
apscheduler>=3.10.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Shared ML infrastructure (cache + prediction logging + Celery)
redis>=5.0.0
celery[redis]>=5.4.0
celery[redis]>=5.4.0
